+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/components/manage-case-flags/manage-case-flags.component.spec.ts
@@ -3,7 +3,7 @@ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { FormGroup, ReactiveFormsModule } from '@angular/forms';
 import { RpxLanguage, RpxTranslationService } from 'rpx-xui-translation';
 import { BehaviorSubject, Observable, of } from 'rxjs';
-import { CaseField } from '../../../../../domain';
+import { CaseField, FieldType } from '../../../../../domain';
 import { MockRpxTranslatePipe } from '../../../../../test/mock-rpx-translate.pipe';
 import { FlagDetail, FlagDetailDisplayWithFormGroupPath, FlagsWithFormGroupPath } from '../../domain';
 import { CaseFlagDisplayContextParameter, CaseFlagFieldState, CaseFlagWizardStepTitle, SelectFlagErrorMessage } from '../../enums';
@@ -14,37 +14,119 @@ describe('ManageCaseFlagsComponent', () => {
   let component: ManageCaseFlagsComponent;
   let fixture: ComponentFixture<ManageCaseFlagsComponent>;
   let mockRpxTranslationService: any;
+  const flag1 = {
+    name: 'Flag 1',
+    flagComment: 'First flag',
+    dateTimeCreated: new Date(),
+    path: [{ id: null, value: 'Reasonable adjustment' }],
+    hearingRelevant: false,
+    flagCode: 'FL1',
+    status: 'Active'
+  };
+  const flag2 = {
+    name: 'Flag 2',
+    flagComment: 'Rose\'s second flag',
+    dateTimeCreated: new Date(),
+    path: [{ id: null, value: 'Reasonable adjustment' }],
+    hearingRelevant: false,
+    flagCode: 'FL2',
+    status: 'Inactive'
+  };
+  const flagX = {
+    name: 'Flag X',
+    flagComment: 'Flag not approved',
+    dateTimeCreated: new Date(),
+    path: [{ id: null, value: 'Reasonable adjustment' }],
+    hearingRelevant: false,
+    flagCode: 'FL1',
+    status: 'Not approved'
+  };
+  const flag3 = {
+    name: 'Flag 3',
+    flagComment: 'First flag',
+    dateTimeCreated: new Date(),
+    path: [{ id: null, value: 'Reasonable adjustment' }],
+    hearingRelevant: false,
+    flagCode: 'FL1',
+    status: 'Active'
+  };
+  const flag4 = {
+    name: 'Flag 4',
+    flagComment: 'Fourth flag',
+    dateTimeCreated: new Date(),
+    path: [
+      { id: null, value: 'Level 1' },
+      { id: null, value: 'Level 2' }
+    ],
+    hearingRelevant: false,
+    flagCode: 'FL1',
+    status: 'Active'
+  };
+  const flag5 = {
+    name: 'Flag 5',
+    name_cy: 'Fflag 5',
+    flagComment: 'Fifth flag',
+    flagComment_cy: 'Fifth flag - Welsh',
+    dateTimeCreated: new Date(),
+    path: [
+      { id: null, value: 'Level 1' }
+    ],
+    hearingRelevant: false,
+    flagCode: 'FL1',
+    status: 'Active',
+    subTypeKey: 'Dummy subtype key',
+    subTypeValue: 'Dummy subtype value',
+    subTypeValue_cy: 'Dummy subtype value - Welsh'
+  };
   const flagsData = [
     {
       flags: {
         partyName: 'Rose Bank',
         details: [
           {
-            id: '1234',
-            name: 'Flag 1',
-            flagComment: 'First flag',
-            dateTimeCreated: new Date(),
-            path: [{ id: null, value: 'Reasonable adjustment' }],
-            hearingRelevant: false,
-            flagCode: 'FL1',
-            status: 'Active'
+            id: '95de3cde-9f1b-468b-863b-8bc29ce7e600',
+            ...flag1
           },
           {
-            id: '2345',
-            name: 'Flag 2',
-            flagComment: 'Rose\'s second flag',
-            dateTimeCreated: new Date(),
-            path: [{ id: null, value: 'Reasonable adjustment' }],
-            hearingRelevant: false,
-            flagCode: 'FL2',
-            status: 'Inactive'
+            id: '34aaf022-19f2-4b03-b78c-6e73fc29e3f2',
+            ...flag2
           }
         ] as FlagDetail[],
         flagsCaseFieldId: 'CaseFlag1'
       },
-      pathToFlagsFormGroup: '',
+      pathToFlagsFormGroup: 'CaseFlag1',
       caseField: {
-        id: 'CaseFlag1'
+        id: 'CaseFlag1',
+        field_type: {
+          id: 'Flags',
+          type: 'Complex'
+        } as FieldType,
+        formatted_value: {
+          partyName: 'Rose Bank',
+          details: [
+            {
+              id: '95de3cde-9f1b-468b-863b-8bc29ce7e600',
+              value: { ...flag1, status: 'Requested' }
+            },
+            {
+              id: '34aaf022-19f2-4b03-b78c-6e73fc29e3f2',
+              value: { ...flag2 }
+            }
+          ]
+        },
+        value: {
+          partyName: 'Rose Bank',
+          details: [
+            {
+              id: '95de3cde-9f1b-468b-863b-8bc29ce7e600',
+              value: { ...flag1 }
+            },
+            {
+              id: '34aaf022-19f2-4b03-b78c-6e73fc29e3f2',
+              value: { ...flag2 }
+            }
+          ]
+        }
       } as CaseField
     },
     {
@@ -52,31 +134,49 @@ describe('ManageCaseFlagsComponent', () => {
         partyName: 'Tom Atin',
         details: [
           {
-            id: '7890',
-            name: 'Flag X',
-            flagComment: 'Flag not approved',
-            dateTimeCreated: new Date(),
-            path: [{ id: null, value: 'Reasonable adjustment' }],
-            hearingRelevant: false,
-            flagCode: 'FL1',
-            status: 'Not approved'
+            id: 'd9348799-1532-444a-b577-cd546e2f58f1',
+            ...flagX
           },
           {
-            id: '3456',
-            name: 'Flag 3',
-            flagComment: 'First flag',
-            dateTimeCreated: new Date(),
-            path: [{ id: null, value: 'Reasonable adjustment' }],
-            hearingRelevant: false,
-            flagCode: 'FL1',
-            status: 'Active'
+            id: '013abc9c-738c-4634-85fa-2910a8cd40a4',
+            ...flag3
           }
         ] as FlagDetail[],
         flagsCaseFieldId: 'CaseFlag2'
       },
-      pathToFlagsFormGroup: '',
+      pathToFlagsFormGroup: 'CaseFlag2',
       caseField: {
-        id: 'CaseFlag2'
+        id: 'CaseFlag2',
+        field_type: {
+          id: 'Flags',
+          type: 'Complex'
+        } as FieldType,
+        formatted_value: {
+          partyName: 'Tom Atin',
+          details: [
+            {
+              id: 'd9348799-1532-444a-b577-cd546e2f58f1',
+              value: { ...flagX }
+            },
+            {
+              id: '013abc9c-738c-4634-85fa-2910a8cd40a4',
+              value: { ...flag3, status: 'Requested' }
+            }
+          ]
+        },
+        value: {
+          partyName: 'Tom Atin',
+          details: [
+            {
+              id: 'd9348799-1532-444a-b577-cd546e2f58f1',
+              value: { ...flagX }
+            },
+            {
+              id: '013abc9c-738c-4634-85fa-2910a8cd40a4',
+              value: { ...flag3 }
+            }
+          ]
+        }
       } as CaseField
     },
     {
@@ -84,24 +184,37 @@ describe('ManageCaseFlagsComponent', () => {
         partyName: '',
         details: [
           {
-            id: '4567',
-            name: 'Flag 4',
-            flagComment: 'Fourth flag',
-            dateTimeCreated: new Date(),
-            path: [
-              { id: null, value: 'Level 1' },
-              { id: null, value: 'Level 2' }
-            ],
-            hearingRelevant: false,
-            flagCode: 'FL1',
-            status: 'Active'
+            id: '573ce187-abae-4121-81d3-53fbba074e62',
+            ...flag4
           }
         ] as FlagDetail[],
         flagsCaseFieldId: 'CaseFlag3'
       },
       pathToFlagsFormGroup: 'caseFlags',
       caseField: {
-        id: 'CaseFlag3'
+        id: 'CaseFlag3',
+        field_type: {
+          id: 'Flags',
+          type: 'Complex'
+        } as FieldType,
+        formatted_value: {
+          partyName: '',
+          details: [
+            {
+              id: '573ce187-abae-4121-81d3-53fbba074e62',
+              value: { ...flag4, status: 'Requested' }
+            }
+          ]
+        },
+        value: {
+          partyName: '',
+          details: [
+            {
+              id: '573ce187-abae-4121-81d3-53fbba074e62',
+              value: { ...flag4 }
+            }
+          ]
+        }
       } as CaseField
     },
     {
@@ -109,28 +222,37 @@ describe('ManageCaseFlagsComponent', () => {
         partyName: '',
         details: [
           {
-            id: '5678',
-            name: 'Flag 5',
-            name_cy: 'Fflag 5',
-            flagComment: 'Fifth flag',
-            flagComment_cy: 'Fifth flag - Welsh',
-            dateTimeCreated: new Date(),
-            path: [
-              { id: null, value: 'Level 1' }
-            ],
-            hearingRelevant: false,
-            flagCode: 'FL1',
-            status: 'Active',
-            subTypeKey: 'Dummy subtype key',
-            subTypeValue: 'Dummy subtype value',
-            subTypeValue_cy: 'Dummy subtype value - Welsh'
+            id: 'aa692b36-9dc0-43de-b5f8-2dba78664376',
+            ...flag5
           }
         ] as FlagDetail[],
         flagsCaseFieldId: 'CaseFlag3'
       },
       pathToFlagsFormGroup: 'caseFlags',
       caseField: {
-        id: 'CaseFlag3'
+        id: 'CaseFlag3',
+        field_type: {
+          id: 'Flags',
+          type: 'Complex'
+        } as FieldType,
+        formatted_value: {
+          partyName: '',
+          details: [
+            {
+              id: 'aa692b36-9dc0-43de-b5f8-2dba78664376',
+              value: { ...flag5, status: 'Requested' }
+            }
+          ]
+        },
+        value: {
+          partyName: '',
+          details: [
+            {
+              id: 'aa692b36-9dc0-43de-b5f8-2dba78664376',
+              value: { ...flag5 }
+            }
+          ]
+        }
       } as CaseField
     }
   ] as FlagsWithFormGroupPath[];
@@ -221,6 +343,7 @@ describe('ManageCaseFlagsComponent', () => {
     expect(displayResult.flagDetailDisplay.flagDetail.name).toEqual(flagDetail.name);
     expect(displayResult.flagDetailDisplay.flagDetail.flagComment).toEqual(flagDetail.flagComment);
     expect(displayResult.flagDetailDisplay.flagDetail.flagCode).toEqual(flagDetail.flagCode);
+    expect(displayResult.originalStatus).toBeUndefined();
     expect(displayResult.pathToFlagsFormGroup).toEqual(flagsInstance.flags.flagsCaseFieldId);
   });
 
@@ -232,21 +355,96 @@ describe('ManageCaseFlagsComponent', () => {
     expect(component.flagsDisplayData[0].flagDetailDisplay.flagDetail.name).toEqual(flagsData[0].flags.details[0].name);
     expect(component.flagsDisplayData[0].flagDetailDisplay.flagDetail.flagComment).toEqual(flagsData[0].flags.details[0].flagComment);
     expect(component.flagsDisplayData[0].flagDetailDisplay.flagDetail.flagCode).toEqual(flagsData[0].flags.details[0].flagCode);
+    expect(component.flagsDisplayData[0].pathToFlagsFormGroup).toEqual(flagsData[0].pathToFlagsFormGroup);
+    expect(component.flagsDisplayData[0].originalStatus).toEqual(flagsData[0].caseField.formatted_value.details[0].value.status);
     // Check correct mapping of the second party's flags
     expect(component.flagsDisplayData[1].flagDetailDisplay.partyName).toEqual(flagsData[1].flags.partyName);
     expect(component.flagsDisplayData[1].flagDetailDisplay.flagDetail.name).toEqual(flagsData[1].flags.details[1].name);
     expect(component.flagsDisplayData[1].flagDetailDisplay.flagDetail.flagComment).toEqual(flagsData[1].flags.details[1].flagComment);
     expect(component.flagsDisplayData[1].flagDetailDisplay.flagDetail.flagCode).toEqual(flagsData[1].flags.details[1].flagCode);
+    expect(component.flagsDisplayData[1].pathToFlagsFormGroup).toEqual(flagsData[1].pathToFlagsFormGroup);
+    expect(component.flagsDisplayData[1].originalStatus).toEqual(flagsData[1].caseField.formatted_value.details[1].value.status);
     // Check correct mapping of the third party's flags
     expect(component.flagsDisplayData[2].flagDetailDisplay.partyName).toEqual(flagsData[2].flags.partyName);
     expect(component.flagsDisplayData[2].flagDetailDisplay.flagDetail.name).toEqual(flagsData[2].flags.details[0].name);
     expect(component.flagsDisplayData[2].flagDetailDisplay.flagDetail.flagComment).toEqual(flagsData[2].flags.details[0].flagComment);
     expect(component.flagsDisplayData[2].flagDetailDisplay.flagDetail.flagCode).toEqual(flagsData[2].flags.details[0].flagCode);
+    expect(component.flagsDisplayData[2].pathToFlagsFormGroup).toEqual(flagsData[2].pathToFlagsFormGroup);
+    expect(component.flagsDisplayData[2].originalStatus).toEqual(flagsData[2].caseField.formatted_value.details[0].value.status);
     // Check correct mapping of the fourth party's flags
     expect(component.flagsDisplayData[3].flagDetailDisplay.partyName).toEqual(flagsData[3].flags.partyName);
     expect(component.flagsDisplayData[3].flagDetailDisplay.flagDetail.name).toEqual(flagsData[3].flags.details[0].name);
     expect(component.flagsDisplayData[3].flagDetailDisplay.flagDetail.flagComment).toEqual(flagsData[3].flags.details[0].flagComment);
     expect(component.flagsDisplayData[3].flagDetailDisplay.flagDetail.flagCode).toEqual(flagsData[3].flags.details[0].flagCode);
+    expect(component.flagsDisplayData[3].pathToFlagsFormGroup).toEqual(flagsData[3].pathToFlagsFormGroup);
+    expect(component.flagsDisplayData[3].originalStatus).toEqual(flagsData[3].caseField.formatted_value.details[0].value.status);
+  });
+
+  it('should map a flag instance for display, where the parent Flags object is a sub-field of a Complex type', () => {
+    const flag1Detail = {
+      id: 'a3e71b29-3f4c-4f60-850c-f002a4d686a8',
+      ...flag1
+    } as FlagDetail;
+    const flagsInstance = {
+      flags: {
+        partyName: 'Ruki Tsunama',
+        details: [
+          flag1Detail,
+          {
+            id: '5fa87240-1856-45a3-9fa7-4dcc8b559396',
+            ...flag2
+          }
+        ] as FlagDetail[],
+        flagsCaseFieldId: 'Witness1'
+      },
+      pathToFlagsFormGroup: 'Witness1.partyFlags',
+      caseField: {
+        id: 'Witness1',
+        field_type: {
+          id: 'Witness',
+          type: 'Complex'
+        } as FieldType,
+        formatted_value: {
+          firstName: 'Ruki',
+          lastName: 'Tsunama',
+          partyFlags: {
+            partyName: 'Ruki Tsunama',
+            details: [
+              {
+                id: 'a3e71b29-3f4c-4f60-850c-f002a4d686a8',
+                value: { ...flag1, status: 'Requested' }
+              },
+              {
+                id: '5fa87240-1856-45a3-9fa7-4dcc8b559396',
+                value: { ...flag2 }
+              }
+            ]
+          }
+        },
+        value: {
+          firstName: 'Ruki',
+          lastName: 'Tsunama',
+          partyFlags: {
+            partyName: 'Ruki Tsunama',
+            details: [
+              {
+                id: 'a3e71b29-3f4c-4f60-850c-f002a4d686a8',
+                value: { ...flag1 }
+              },
+              {
+                id: '5fa87240-1856-45a3-9fa7-4dcc8b559396',
+                value: { ...flag2 }
+              }
+            ]
+          }
+        }
+      } as CaseField
+    } as FlagsWithFormGroupPath;
+    const flagDetailMappedForDisplay = component.mapFlagDetailForDisplay(flag1Detail, flagsInstance);
+    expect(flagDetailMappedForDisplay.flagDetailDisplay.partyName).toEqual(flagsInstance.flags.partyName);
+    expect(flagDetailMappedForDisplay.flagDetailDisplay.flagDetail).toEqual(flag1Detail);
+    expect(flagDetailMappedForDisplay.pathToFlagsFormGroup).toEqual(flagsInstance.pathToFlagsFormGroup);
+    expect(flagDetailMappedForDisplay.originalStatus).toEqual(flagsInstance.caseField.formatted_value.partyFlags.details[0].value.status);
   });
 
   it('should emit to parent with the selected party and flag details if the validation succeeds', () => {
@@ -265,9 +463,10 @@ describe('ManageCaseFlagsComponent', () => {
           flagDetail: flagsData[1].flags.details[1],
           flagsCaseFieldId: flagsData[1].flags.flagsCaseFieldId
         },
-        pathToFlagsFormGroup: '',
+        pathToFlagsFormGroup: flagsData[1].pathToFlagsFormGroup,
         caseField: flagsData[1].caseField,
-        roleOnCase: undefined
+        roleOnCase: undefined,
+        originalStatus: flagsData[1].caseField.formatted_value.details[1].value.status
       } as FlagDetailDisplayWithFormGroupPath
     });
     expect(component.errorMessages.length).toBe(0);