+++ projects/ccd-case-ui-toolkit/src/lib/shared/components/palette/case-flag/domain/case-flag.model.ts
@@ -55,7 +55,8 @@ export interface FlagsWithFormGroupPath {
 }
 
 /**
- * Wrapper interface for FlagDetailDisplay that adds the path to the corresponding FormGroup, and the CaseField
+ * Wrapper interface for FlagDetailDisplay that adds the path to the corresponding FormGroup, the CaseField, and the
+ * original flag status
  */
 export interface FlagDetailDisplayWithFormGroupPath {
   flagDetailDisplay: FlagDetailDisplay;
@@ -63,4 +64,5 @@ export interface FlagDetailDisplayWithFormGroupPath {
   caseField: CaseField;
   roleOnCase?: string;
   label?: string;
+  originalStatus?: string;
 }